Refine synthetic resource binding API layout
