+++ src/cuda/cuda-device.cpp
@@ -13,6 +13,7 @@
 #include "cuda-shader-table.h"
 #include "cuda-utils.h"
 #include "cuda-heap.h"
+#include "../synthetic-resource-bindings.h"
 
 namespace rhi::cuda {
