+++ src/cuda/cuda-shader-object-layout.h
@@ -2,9 +2,12 @@
 
 #include "cuda-base.h"
 
+#include <algorithm>
+
 namespace rhi {
 struct SyntheticResourceBindingRecord;
-}
+struct SyntheticBindingLocation;
+} // namespace rhi
 
 namespace rhi::cuda {
 
@@ -46,14 +49,17 @@ public:
 
     uint32_t m_slotCount = 0;
     uint32_t m_subObjectCount = 0;
-    size_t m_minUniformBufferSize = 0;
+    size_t m_uniformBufferSize = 0;
 
     ShaderObjectLayoutImpl(Device* device, slang::ISession* session, slang::TypeLayoutReflection* layout);
 
     // ShaderObjectLayout interface
     virtual uint32_t getSlotCount() const override { return m_slotCount; }
     virtual uint32_t getSubObjectCount() const override { return m_subObjectCount; }
-    virtual size_t getMinUniformBufferSize() const override { return m_minUniformBufferSize; }
+    virtual size_t getUniformBufferSize() const override
+    {
+        return std::max(Super::getUniformBufferSize(), m_uniformBufferSize);
+    }
 
     virtual uint32_t getBindingRangeCount() const override { return m_bindingRanges.size(); }
     virtual const BindingRangeInfo& getBindingRange(uint32_t index) const override { return m_bindingRanges[index]; }
