+++ src/cuda/cuda-shader-object.cpp
@@ -137,13 +137,7 @@ Result BindingDataBuilder::writeObjectData(
     ObjectData& outData
 )
 {
-    size_t size = specializedLayout->getElementTypeLayout()->getSize();
-    // Widen the allocation to satisfy any synthesized resources whose
-    // uniform offset lies outside the Slang element-type layout. Mirrors
-    // the sizing performed by ShaderObject::init for `m_data`.
-    const size_t minUniformSize = specializedLayout->getMinUniformBufferSize();
-    if (minUniformSize > size)
-        size = minUniformSize;
+    size_t size = specializedLayout->getUniformBufferSize();
 
     ConstantBufferPool::Allocation allocation;
     SLANG_RETURN_ON_FAIL(m_constantBufferPool->allocate(size, memType, allocation));
